feat: rename ui prompting examples package
rename ui prompting examples package
