+++ examples/ui-prompting/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@sap-ux/prompting-ui",
+    "name": "@sap-ux-private/ui-prompting-examples",
     "version": "0.1.7",
     "description": "This project contains UI storybook stories with exampleS with prompt ui and FPM based building blocks.",
     "license": "Apache-2.0",
