+++ examples/ui-prompting/scripts/index.js
