+++ examples/ui-prompting/src/BuildingBlock.tsx
