+++ examples/ui-prompting/src/Chart.story.tsx
