+++ examples/ui-prompting/src/Custom.story.tsx
