+++ examples/ui-prompting/src/FilterBar.story.tsx
