+++ examples/ui-prompting/src/Input.story.tsx
