+++ examples/ui-prompting/src/Table.story.tsx
