+++ examples/ui-prompting/src/addons/project/actions.ts
