+++ examples/ui-prompting/src/addons/project/index.ts
