+++ examples/ui-prompting/src/addons/register.ts
