+++ examples/ui-prompting/src/addons/types.ts
