+++ examples/ui-prompting/src/backend/api.ts
