+++ examples/ui-prompting/src/backend/i18nBundle.ts
