+++ examples/ui-prompting/src/utils/communication.ts
