+++ examples/ui-prompting/src/utils/hooks.ts
