+++ examples/ui-prompting/src/utils/index.ts
