+++ examples/ui-prompting/src/utils/types.ts
