+++ examples/ui-prompting/tsconfig.json
